Update retweet functionality
Added a drop menu when clicking the retweet button, for the user to
choose between a simple retweet or a quote tweet.
Also added a check in the case of a user setting its profile to private
'right before' someone retweets their tweet, so if a user tries to do it
while the menu still appears, they can't actually retweet.
Updated the coloring of the numbers and updated symbols to reflect when a tweet
was retweeted and when a user cannot retweet due to the author having a
private account