Set no-cache headers
If for example, the user clicks on a tweet and then from the tweet's
show view, retweets it and then goes back to the previous page, if the
page is cached, it won't show the new retweet. So, to prevent that, I
set the no-cache headers, so that the new tweets will always be loaded.

I realize that this might come at the cost of performance, but I feel
like at this moment it doesn't penalize much. Plus, it is annoying to
manually refresh the page in order to show the new tweets.