Set up broadcasts for retweets and update feature
Include broadcasting for the retweets, in a similar way to the likes,
where the counters are updated globally, also tackling the problem where
a user tries to "toy" with the retweets on separate tabs.

Also include a check and prevention of the issue for the case if the
author sets their account to private and a user that sees the authors
tweet still has the not updated privated version of the button.