Update Comments layout to be Reddit-like
Introduce depth, height and root to tweets and comments.
This is because I want to allow users to reply to the original tweet and
the replies/comments under it.
As I don't want the end result to be confusing to the user in terms of
UI, like it is on Twitter, where the comments are streamlined with no
clear distinction of who replies to what, I decided to go with a more
nested like structure, and those CS concepts had to be introduced (at
least I believe it is a good way of achieving the desired result), to
achieve a tree-like structure of the comments, with clean organization
and possibility to organize the layout better, so a single page won't
have infinite comments. I will still tinker with it, as it is not the
final version