Fix error with broadcasting messages to all users
The broadcasting was setup in such a way that it was targeting a user
through their GID, meaning that all received/sent messages to
conversations where they participated was being broadcasted to everyone.
This fix makes sure that the broadcast is targeting the user AND the
conversation, so each message is only broadcasted to the respective
conversations and users that take part in it.