"Correct the behavior when sharing posts in dms"
When users send just the post, without anything else to add, it just
renders/shows the post. When the user adds some small comment to the
message, it renders the post and a separate 'comment', without
cluttering the message with the url. Now, when the user sends multiple
urls in the same message, it is maybe useful to keep track of the order
of the posts/urls, so in that case, the posts themselves are embedded as
previously, but the comment message shows the full message with urls and
text. The only detail is that the urls are truncated.