Filter URL's, validate and HyperLink them
Not all url's are valid. There are specific rules. But some of those rules, imo
can be bent for clarity and readability.

There are specific TLD's that are not considered valid. I introduced a
filter/validation check when creating a tweet/message, that checks for
it and responds accordingly. Also, hyperlink only the relevant parts of
the urls, which are considered valid. Not perfect, but i will go as to
say that it works for the majority of the cases for the general
population (assuming that this population won't be trying to break and
limit test this validation/rule).