Chrome users will have no 'smooth' transition
Because of a weird - seemingly unfixed bug ([almost 10 years, maybe even more](https://stackoverflow.com/questions/22069877/css-transitions-strange-unwanted-delay-in-webkit-browsers-chrome-and-safari) or this is the way chrome
operates, which imo is not the best of the possible ways, firefox does this much better.

So, instead of spending more time to find workarounds (which I've tried, believe me) that make the code really really ugly, and still don't work perfectly, as I have to account for every child element inheriting or not inheriting when it is or isn't needed, the simple fix is just take that hassle whatsoever from chromium browsers. If the user is on a chrome browser, basically all of their color related transitions have a duration of 0s, which eliminate the ugly and tedious 'cascading' color effect. On firefox it works pretty nice though.