Clicking on a searched message goes to that msg
Also added other features and fixes related to messages, that I didn't
think needed their separate commit.

So, if a user searches up some message/string, and it is found, they can
click on it and the conversation and message is automatically scrolled into their view.

Modified some css related to the messages

Fixed up the search method in the msg controller, as it was looking
through all of the messages. Now looks only through the messages the
user sent/received.

Got rid of useless files